// src/components/atoms/button/Button.types.ts
import { ButtonHTMLAttributes, ReactNode } from "react";

export type ButtonVariant = 'primary' | 'secondary' | 'outline' | 'ghost' ;

export type ButtonSize = 'sm' | 'md' | 'lg' | 'xl' ;

export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
    variant?: ButtonVariant;
    size?: ButtonSize;
    children: ReactNode;
    isLoading?: boolean;
    fullWidth?: boolean;
}